Removed Position interface
